:truck: move `internal/logstorage` to `internal/data/logstorage`
