+++ api/main.go
@@ -12,7 +12,7 @@ import (
 
 	"link-society.com/flowg/internal"
 	"link-society.com/flowg/internal/data/auth"
-	"link-society.com/flowg/internal/logstorage"
+	"link-society.com/flowg/internal/data/logstorage"
 	"link-society.com/flowg/internal/pipelines"
 )
