+++ api/query_logs.go
@@ -10,8 +10,8 @@ import (
 	"github.com/swaggest/usecase/status"
 
 	"link-society.com/flowg/internal/data/auth"
+	"link-society.com/flowg/internal/data/logstorage"
 	"link-society.com/flowg/internal/filterdsl"
-	"link-society.com/flowg/internal/logstorage"
 )
 
 type QueryStreamRequest struct {
