+++ cmd/flowg/serve.go
@@ -16,7 +16,7 @@ import (
 
 	"link-society.com/flowg/internal/app/logging"
 	"link-society.com/flowg/internal/data/auth"
-	"link-society.com/flowg/internal/logstorage"
+	"link-society.com/flowg/internal/data/logstorage"
 	"link-society.com/flowg/internal/pipelines"
 
 	"link-society.com/flowg/api"
