+++ internal/data/logstorage/entry.go
