+++ internal/data/logstorage/errors.go
