+++ web/apps/streams/controllers/default_page.go
@@ -8,7 +8,7 @@ import (
 	"github.com/a-h/templ"
 
 	"link-society.com/flowg/internal/data/auth"
-	"link-society.com/flowg/internal/logstorage"
+	"link-society.com/flowg/internal/data/logstorage"
 
 	"link-society.com/flowg/web/apps/streams/templates/views"
 )
