+++ web/apps/streams/controllers/stream_page.go
@@ -12,8 +12,8 @@ import (
 	"github.com/a-h/templ"
 
 	"link-society.com/flowg/internal/data/auth"
+	"link-society.com/flowg/internal/data/logstorage"
 	"link-society.com/flowg/internal/filterdsl"
-	"link-society.com/flowg/internal/logstorage"
 
 	"link-society.com/flowg/web/apps/streams/templates/components"
 	"link-society.com/flowg/web/apps/streams/templates/views"
