+++ web/apps/streams/main.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"link-society.com/flowg/internal/data/auth"
-	"link-society.com/flowg/internal/logstorage"
+	"link-society.com/flowg/internal/data/logstorage"
 
 	"link-society.com/flowg/web/apps/streams/controllers"
 )
