+++ web/apps/streams/templates/components/viewer.templ
@@ -3,7 +3,7 @@ package components
 import (
   "time"
 
-	"link-society.com/flowg/internal/logstorage"
+	"link-society.com/flowg/internal/data/logstorage"
 )
 
 type ViewerProps struct {
