+++ web/apps/streams/templates/views/page.templ
@@ -4,7 +4,7 @@ import (
   "time"
 
 	"link-society.com/flowg/internal/data/auth"
-	"link-society.com/flowg/internal/logstorage"
+	"link-society.com/flowg/internal/data/logstorage"
 
   "link-society.com/flowg/web/apps/streams/templates/components"
   "link-society.com/flowg/web/templates/layouts"
